+++ resources/views/livewire/pages/auth/register.blade.php
@@ -5,14 +5,45 @@ use App\Providers\RouteServiceProvider;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Str;
 use Illuminate\Validation\Rules;
 use Livewire\Attributes\Layout;
 use Livewire\Volt\Component;
+use ZxcvbnPhp\Zxcvbn;
 
 new #[Layout('layouts.guest')] class extends Component
 {
     public string $name = '';
     public string $email = '';
+    public string $password = '';
+    public string $password_confirmation = '';
+    public int $strengthScore = 0;
+    public array $strengthLevels = [
+        1 => 'Weak',
+        2 => 'Fair',
+        3 => 'Good',
+        4 => 'Strong',
+    ];
+
+    public function generatePassword(): void
+    {
+        $password = Str::password(12);
+
+        $this->password = $password;
+        $this->password_confirmation = $password;
+
+        $this->updateStrengthScore();
+    }
+
+    public function updatedPassword($value)
+    {
+        $this->updateStrengthScore();
+    }
+
+    protected function updateStrengthScore(): void
+    {
+        $this->strengthScore = (new Zxcvbn())->passwordStrength($this->password)['score'];
+    }
 
     /**
      * Handle an incoming registration request.
@@ -40,18 +71,18 @@ new #[Layout('layouts.guest')] class extends Component
         <!-- Name -->
         <div>
             <x-input-label for="name" :value="__('Name')" />
-            <x-text-input wire:model="name" id="name" class="block w-full mt-1" type="text" name="name" required autofocus autocomplete="name" />
+            <x-text-input wire:model="name" id="name" class="block w-full mt-1" type="text" name="name" autofocus autocomplete="name" />
             <x-input-error :messages="$errors->get('name')" class="mt-2" />
         </div>
 
         <!-- Email Address -->
         <div class="mt-4">
             <x-input-label for="email" :value="__('Email')" />
-            <x-text-input wire:model="email" id="email" class="block w-full mt-1" type="email" name="email" required autocomplete="username" />
+            <x-text-input wire:model="email" id="email" class="block w-full mt-1" type="email" name="email" autocomplete="username" />
             <x-input-error :messages="$errors->get('email')" class="mt-2" />
         </div>
 
-        @livewire('register-passwords')
+        <x-register-passwords :$strengthLevels :$strengthScore />
 
         <div class="flex items-center justify-end mt-4">
             <a class="text-sm text-gray-600 underline rounded-md hover:text-gray-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500" href="{{ route('login') }}" wire:navigate>
